requirements.txt
pytelegrambotapi==4.10.0
openai==0.26.5
PyYAML==6.0
